requirements.txt
asyncio>=3.4.3
sanic>=25.3.0
dotenv>=0.9.9
pydantic>=2.11.4
shortuuid>=1.0.13
arize-phoenix-otel>=0.12.1